+++ joulescope_ui/widgets/waveform/signal.py
@@ -537,7 +537,7 @@ class Signal(QtCore.QObject):
 
     @QtCore.Slot(object, object)
     def _on_y_range_changed(self, vb, y_range):
-        self.log.info('_on_y_range_changed(%s, %s)', self.name, y_range)
+        self.log.debug('_on_y_range_changed(%s, %s)', self.name, y_range)
         for m in self._markers_single.values():
             m.move(self.vb)
         for m in self._markers_dual.values():